[🛠️ REFACTOR] fetchRoomPlayers 함수 모듈화
